+++ NEWS.md
@@ -1,6 +1,6 @@
 # dtutils (development version)
 
-* Added `data.frame` method for `write_tsv()` (#1).
+* Added `data.frame` and `matrix` method for `write_tsv()` (#1, #2).
 * Improved quality control and documentation.
 
 # dtutils 0.1.0